requirement.txt
pandas
numpy
pytorch
spacy
PIL
shutil
-e .